refactor: change MultiVideoSelect file location
